Fix potential (differs per system) cyclic dependency problem in Sorta
